Keep calm and carry on
See GitHub issue #41.

Don't panic when calculating a pdf, cdf or pmf whenever the input is a
value that the distribution cannot attain. The pdf, pmf and cdf are
still defined at those inputs and we should return the correct values.

This commit also adds a few tests to each distribution to test that
pdf(-inf)=0, pdf(inf)=0, cdf(-inf)=0 and cdf(inf)=1. It also uses simple
numerical integration to test for continuous distributions that the
integral of the pdf is approximately equal to the cdf, and for discrete
distributions that the sum of the pmf is equal to the cdf.