Update to rand 0.5 (#89)
* Update to rand 0.5

This required implementing `Distribution` in terms of Rand's
`Distribution`, because it is more general.

* Apply rustfmt

* Fix deprecation warning

* Apply more recent rustfmt

* Remove `Distribution` and `WeakRngDistribution`

It is superseded by `rand::distributions::Distribution`.

`Univariate` was changed to no longer require `Distribution`.

Fixes #88.

* Don't alias Rand's `Distribution`

* Restore information about sampling algorithms

Some of the comments deleted before contained information on the
implementation of sampling. This information has been restored and put
to appropriate places.