test: confirm chi mean fails at dof > ~340
This test shows that the mean of the chi distribution produces NaN at above approximately 340.0 degrees of freedom.
When the dof exceeds this, the gamma function is called with an argument greater than approximately 170.0 producing a result with exceeds the maximum normal value for a f64.
The equivalent limit for an f32 is approx half at a chi dof of 170.0 and a gamma argument of 85.0.
As this method is key to distinguishing between stationary/nonstationary gaussian process, which can be high dimensional in many applications, this should be fixed.