fix: impl high dof approximation for Chi mean and variance
This implements the a large n approximation for the mean of the Chi distribution, preventing NaN from being returned for values above ~340.
This extends the series approximation provided on Wikipedia with enough terms so that the accuracy is roughly equivalent to the numerical error of the existing implementation for values near the threshold.
As the relative error is bounded by O(1/n^4) it cannot be used for small values, where the current implementation is better.
The added test checks the new value against the value produced by the old implementation near the threshold.