refactor!: decouple distribution traits from rand
This changes the behavior of Cauchy::mean() and Cauchy::variance()
from the old default implementation (based on RNG sampling) to simply
returning `None`. Both mean and variance are undefined for the cauchy
distribution.

This also changes the trait signature of both Distribution<T> and
DiscreteDistribution<T> to not be subtraits of
rand::distribution::Distribution<T>. Our distributions do still
implement this trait iff the `rand` feature is enabled.