refactor!: Chi: Use u64 for `freedom`
Degrees of freedom `k` is usually defined as a positive
integer. Storing `k` as a NonZeroU64 avoids all degenerate cases
from the start (k < 0, k == 0, k == NaN), and the overhead of
casting to an f64 for some operations should be minimal.

NonZeroU64 also allows memory layout optimizations (similar to
the nullptr optimization).